Fix test structure and file moves
